Create README.md
This project aims at predicting anxiety and depression cases among high school students in Kenya. It spans about four counties in particular. Nairobi, Kiambu, Makueni, and Machakos. It aims to uncover the underlying factors contributing
It provides initiatives to be taken by the stakeholders to ensure at-risk students are taken care of appropriately and promptly.